Move mixin class to mixin director. Add defaults for meta and path, and move meta back to model.
