+++ tethysext/atcore/models/file_database/file_collection.py
@@ -11,10 +11,11 @@ from typing import Generator
 import uuid
 
 from sqlalchemy import Column, ForeignKey
+from sqlalchemy.dialects.postgresql import JSON
 from sqlalchemy.orm import relationship
 
+from tethysext.atcore.mixins.meta_mixin import MetaMixin
 from tethysext.atcore.models.file_database.base import FileDatabaseBase
-from tethysext.atcore.models.file_database.meta_mixin import MetaMixin
 from tethysext.atcore.models.types import GUID
 
 
@@ -24,6 +25,7 @@ class FileCollection(MetaMixin, FileDatabaseBase):
 
     id = Column('id', GUID, primary_key=True, default=uuid.uuid4)
     file_database_id = Column('file_database_id', GUID, ForeignKey('file_databases.id'))
+    meta = Column('metadata', JSON)
 
     database = relationship("FileDatabase", back_populates="collections")
