+++ tethysext/atcore/models/file_database/file_database.py
@@ -10,10 +10,11 @@ import os
 import uuid
 
 from sqlalchemy import Column, String
+from sqlalchemy.dialects.postgresql import JSON
 from sqlalchemy.orm import relationship
 
+from tethysext.atcore.mixins.meta_mixin import MetaMixin
 from tethysext.atcore.models.file_database.base import FileDatabaseBase
-from tethysext.atcore.models.file_database.meta_mixin import MetaMixin
 from tethysext.atcore.models.types import GUID
 
 
@@ -23,6 +24,7 @@ class FileDatabase(MetaMixin, FileDatabaseBase):
 
     id = Column('id', GUID, primary_key=True, default=uuid.uuid4)
     root_directory = Column('root_directory', String)
+    meta = Column('metadata', JSON)
 
     collections = relationship("FileCollection", back_populates="database")
